Refactor: Move DAO implementations to 'impl' package for better separation of concerns
