+++ src/test/java/com/napier/sem/dao/CityDAOTest.java
@@ -1,6 +1,6 @@
 package com.napier.sem.dao;
 
-import com.napier.sem.models.City;
+import com.napier.sem.dao.impl.CityDAO;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
@@ -11,9 +11,7 @@ import java.sql.Connection;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
 import java.sql.SQLException;
-import java.util.List;
 
-import static org.junit.jupiter.api.Assertions.*;
 import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.Mockito.*;
