+++ src/test/java/com/napier/sem/dao/CountryDAOTest.java
@@ -1,5 +1,6 @@
 package com.napier.sem.dao;
 
+import com.napier.sem.dao.impl.CountryDAO;
 import com.napier.sem.models.Country;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
